move into proposal and create a default set of params
